requirements.txt
openai>=1.0
anthropic>=0.30
google-genai>=1.0
python-dotenv>=1.0
pandas>=2.0
